+++ keeweb/templates/manifest.appcache.php
@@ -1,0 +1,10 @@
+CACHE MANIFEST
+
+# YYYY-MM-DD:v0.0.0
+# updmin:v0.0.0
+
+CACHE:
+/index.php/apps/keeweb
+
+NETWORK:
+*