+++ keeweb/CHANGELOG.md
@@ -1,2 +1,5 @@
-owncloud-keeweb (0.1.0)
+nextcloud-keeweb (0.2.0)
+* Fix cache manifest
+
+nextcloud-keeweb (0.1.0)
 * Initial release